ADM-359 refactor
