+++ frontend/src/components/Common/Buttons.ts
@@ -1,0 +1,14 @@
+import Button from '@mui/material/Button'
+import { styled } from '@mui/material/styles'
+
+export const VerifyButton = styled(Button)({
+  width: '3rem',
+  fontSize: '0.8rem',
+  fontWeight: '550',
+})
+export const ResetButton = styled(Button)({
+  width: '3rem',
+  fontSize: '0.8rem',
+  fontWeight: '550',
+  color: '#f44336',
+})
