+++ frontend/src/components/ErrorNotification/index.tsx
