+++ frontend/src/components/ErrorNotification/style.tsx
