+++ frontend/src/theme.ts
@@ -1,7 +1,6 @@
-import { createTheme, styled } from '@mui/material/styles'
+import { createTheme } from '@mui/material/styles'
 import { indigo } from '@mui/material/colors'
 import { FIVE_HUNDRED } from '@src/constants'
-import { Button } from '@mui/material'
 
 declare module '@mui/material/styles' {
   interface Theme {
@@ -44,16 +43,4 @@ export const theme = createTheme({
       textTransform: 'none',
     },
   },
-})
-
-export const VerifyButton = styled(Button)({
-  width: '3rem',
-  fontSize: '0.8rem',
-  fontWeight: '550',
-})
-export const ResetButton = styled(Button)({
-  width: '3rem',
-  fontSize: '0.8rem',
-  fontWeight: '550',
-  color: '#f44336',
 })
